f-metadata@2.6.0 - Change initialisation to configuration and run against window.appboy
