+++ packages/f-metadata/src/utils/noop.js
@@ -1,0 +1,3 @@
+const noop = () => {};
+
+export default noop;
